+++ CHANGES.rst
@@ -1,5 +1,10 @@
 .. currentmodule:: click
 
+Unreleased
+
+-   Fix handling of ``flag_value`` when ``is_flag=False`` to allow such options to be
+    used without an explicit value. :issue:`3084`
+
 Version 8.3.1
 --------------